Seratan-Sakedap/requirements.txt
gradio
pypdf
pdf2docx
PyMuPDF
Pillow